commands split into subfiles, custom command decorator, refactoring code
