+++ mantis/commands/__init__.py
@@ -1,0 +1,28 @@
+"""Mantis CLI command modules."""
+from mantis.commands import (
+    core,
+    images,
+    containers,
+    compose,
+    services,
+    crypto,
+    config,
+    volumes,
+    django,
+    postgres,
+    nginx,
+)
+
+__all__ = [
+    'core',
+    'images',
+    'containers',
+    'compose',
+    'services',
+    'crypto',
+    'config',
+    'volumes',
+    'django',
+    'postgres',
+    'nginx',
+]
