+++ mantis/commands/compose.py
@@ -1,0 +1,30 @@
+"""Compose commands: up, down, run."""
+from typing import Optional, List
+
+import typer
+
+from mantis.app import command, state, join_args
+
+
+@command(panel="Compose")
+def up(
+    params: Optional[List[str]] = typer.Argument(None, help="Compose up parameters"),
+):
+    """Calls compose up"""
+    state.up(join_args(params))
+
+
+@command(panel="Compose")
+def down(
+    params: Optional[List[str]] = typer.Argument(None, help="Compose down parameters"),
+):
+    """Calls compose down"""
+    state.down(join_args(params))
+
+
+@command(name="run", panel="Compose")
+def run_cmd(
+    params: List[str] = typer.Argument(..., help="Compose run parameters"),
+):
+    """Calls compose run with params"""
+    state.run(join_args(params))
